Adapting to latest changes in Biojava

+++ eppic-cli/src/main/java/eppic/ChainEvolContext.java
@@ -720,7 +720,7 @@ public class ChainEvolContext implements Serializable {
 			
 			if (queryPos!=-1) { 
 				for (Atom atom:residue.getAtoms()) {
-					atom.setTempFactor(conservationScores.get(queryPos-1));
+					atom.setTempFactor(conservationScores.get(queryPos-1).floatValue());
 				}	
 			} else {
 				// when no entropy info is available for a residue we still want to assign a value for it
@@ -731,7 +731,7 @@ public class ChainEvolContext implements Serializable {
 				LOGGER.info("Residue {} ({}) of chain {} has no entropy value associated to it, will set its b-factor to max entropy ({})",
 						residue.getResidueNumber().toString(), residue.getPDBName(), residue.getChainId(), maxEntropy);
 				for (Atom atom:residue.getAtoms()) {
-					atom.setTempFactor(maxEntropy);
+					atom.setTempFactor((float)maxEntropy);
 				}	
 			}
 		}
